refactor: Theme 적용 방식 변경 (#7)
