+++ src/pages/Main.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import Button from '../components/atomics/Button';
+import Button from '../components/common/Button';
 import { css } from '@emotion/react';
 
 const Main = () => {
