+++ src/stories/IconText.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import IconText from '../components/molecules/IconText';
+import IconText from '../components/SignUp/IconText';
 import React from 'react';
 import { MdKeyboardArrowLeft } from 'react-icons/md';
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
